Move aggregations to compile directory
