+++ site/frontend/src/pages/compare/compile/compile-page.vue
@@ -2,7 +2,7 @@
 import QuickLinks from "./quick-links.vue";
 import Filters from "./filters.vue";
 import OverallSummary from "../summary/overall-summary.vue";
-import Aggregations from "../summary/aggregations.vue";
+import Aggregations from "./aggregations.vue";
 import Benchmarks from "../benchmarks/benchmarks.vue";
 import {CompareResponse, CompareSelector} from "../types";
 import {computed, ref} from "vue";
